little change + code cleanup

+++ RubixMatrix/Dtype.h
@@ -15,7 +15,7 @@ namespace Rubix
 	class Dtype
 	{
 		std::string name_;
-		uint64_t physsize_; // logical size in byte of a single value of type T / Dtype.
+		uint16_t physsize_; // logical size in byte of a single value of type T / Dtype.
 
 		public:
 		__declspec(deprecated("Not Implemented"))
